:recycle: Simplify attachment creation and resuse code between child services.
