+++ api/src/services/travel-desk-travel-requests/passenger-name-record-document/index.ts
@@ -1,1 +1,1 @@
-export { CreateService } from "./create-service"
+export { UpsertService } from "./upsert-service"
